moved tictactoe logic into a separate .dart file
