+++ docs/template-method/script.js.deps
@@ -1,6 +1,7 @@
 file:///C:/tools/dart-sdk/lib/_internal/dart2js_platform.dill
 file:///C:/tools/dart-sdk/lib/libraries.json
 file:///D:/GitHub/patterns/docs/template-method/script.dart
+file:///D:/GitHub/patterns/docs/template-method/tictactoe.dart
 org-dartlang-sdk:///lib/_http/crypto.dart
 org-dartlang-sdk:///lib/_http/embedder_config.dart
 org-dartlang-sdk:///lib/_http/http.dart
